requirements-build.txt
PyQt6>=6.6.0
google-generativeai>=0.8.0
pyaudiowpatch>=0.2.12
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
pyinstaller>=6.0.0
pyinstaller-hooks-contrib>=2024.0
